Disable drdr runs on `meta/web'.
The problem is that it spits out a warning message on stderr whenever it
can't find a git repository to extract release announcements (from
annotated tags).  Resolving this will require some way to (a) tell drdr
to ignore such output, (b) make it run the code with a $GIT_DIR
environment variable set to the git tree, or (c) make it set some
environment variable so the web build will avoid such messages.  (I
prefer the first most, and the last least.)